refactor: updates namespace and classes
