+++ src/main/java/com/example/todos/TodoApplication.java
@@ -1,13 +1,13 @@
-package com.example.demo;
+package com.example.todos;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class DemoApplication {
+public class TodoApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
+		SpringApplication.run(TodoApplication.class, args);
 	}
 
 }
