+++ src/main/java/com/example/todos/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.example.demo.exception;
+package com.example.todos.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
